serial: msm_geni_serial: reconfigure baud rate and port setup
For HSUART we are doing the port setup only once during probe
and if system is entering into deep sleep without port close,
after the deep sleep exit QUP HW is not retaining register
configuration made as part of port setup and set_termios.

Added changes in the deep sleep exit path to do the port setup
and reconfigure uart to the previous baud rate.

Change-Id: Ieb7a1e5fd624de8985be19ff47f579c3d5e92e41
Signed-off-by: Anvesh Salveru <quic_asalveru@quicinc.com>