usb: phy: Prevent regulator disable when eud is active
USB is entering low power mode while target bootup with
cable connected which is resulting in turning off LDOs/clocks
which leads to phy power down. This leads to USB functionality
and EUD not working as PHY won't be re-initialized if EUD is
active.

This is happpening due to VBUS notifcation comes later after
EUD notification and with eud notification, we are checking
for cable connected which fails as vbus_active is not set
at that time and we go peripheral suspend and suspends the
phy.

To make sure not turning off LDOS/clocks while cable is
connected and eud is active as part of set_suspend(0),
we are using this dpdm_enable flag to know cable status
and not turning off.

On some targets 3.3V LDO which acts as EUD power up (which in
turn reset the USB PHY) is shared with EMMC so that it
won't be turned off even though we remove our vote as
part of disconnect. so power up this regulator is actually
not resetting the PHY next time when cable is connected
after disconnect due to which phy cannot resume.

So we explicitly bring it out of power down state
by writing to POWER DOWN register.

Change-Id: Ib142a24a53944448292bbde1723e03030ddaa085
Signed-off-by: AKASH KUMAR <quic_akakum@quicinc.com>