qcom: power: smblite: Move requesting of batt data to complete cb
There can be race condition where-in glink comm driver resume
is called after the resume callback of charger driver. This
may lead to device crash because charger driver requests battery
data when glink is not yet up.

Move requesting of battery data to complete callback of pm_ops
which is called after all the drivers resume callback is completed
to prevent device crash.

Change-Id: I1fffed255024d35d30cc16aadf133fb3a3c5eeb3
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>